Revert "BACKPORT: kasan: make kasan_cache_create() work with 32-bit slab cache sizes"
This reverts commit 57a04e68ea033f40c8c4706f98d34bf0541d09c3.

+++ mm/slab.c
@@ -2013,7 +2013,7 @@ __kmem_cache_create (struct kmem_cache *cachep, unsigned long flags)
 	size_t ralign = BYTES_PER_WORD;
 	gfp_t gfp;
 	int err;
-	unsigned int size = cachep->size;
+	size_t size = cachep->size;
 
 #if DEBUG
 #if FORCED_DEBUG
